requirements.txt
coloredlogs==15.0.1
numpy==1.25.1
Pillow==10.0.0
pure_python_adb==0.3.0.dev0
PyAutoGUI==0.9.54
pynput==1.7.6
rich==13.4.2
schedule==1.2.0
toml==0.10.2
